// src/sunspec/models/MeterModel203Register.ts
/**
 * Register offsets within the data section of
 * SunSpec three-phase meter model 203.
 *
 * The offsets are relative to the first data register,
 * immediately after model ID and model length.
 */
export enum MeterModel203Register {

  CURRENT = 0,

  CURRENT_PHASE_A = 1,

  CURRENT_PHASE_B = 2,

  CURRENT_PHASE_C = 3,

  CURRENT_SCALE_FACTOR = 4,

  VOLTAGE_LINE_NEUTRAL = 5,

  VOLTAGE_PHASE_AN = 6,

  VOLTAGE_PHASE_BN = 7,

  VOLTAGE_PHASE_CN = 8,

  VOLTAGE_LINE_LINE = 9,

  VOLTAGE_PHASE_AB = 10,

  VOLTAGE_PHASE_BC = 11,

  VOLTAGE_PHASE_CA = 12,

  VOLTAGE_SCALE_FACTOR = 13,

  FREQUENCY = 14,

  FREQUENCY_SCALE_FACTOR = 15,

  ACTIVE_POWER = 16,

  ACTIVE_POWER_PHASE_A = 17,

  ACTIVE_POWER_PHASE_B = 18,

  ACTIVE_POWER_PHASE_C = 19,

  ACTIVE_POWER_SCALE_FACTOR = 20,

  APPARENT_POWER = 21,

  APPARENT_POWER_PHASE_A = 22,

  APPARENT_POWER_PHASE_B = 23,

  APPARENT_POWER_PHASE_C = 24,

  APPARENT_POWER_SCALE_FACTOR = 25,

  REACTIVE_POWER = 26,

  REACTIVE_POWER_PHASE_A = 27,

  REACTIVE_POWER_PHASE_B = 28,

  REACTIVE_POWER_PHASE_C = 29,

  REACTIVE_POWER_SCALE_FACTOR = 30,

  POWER_FACTOR = 31,

  POWER_FACTOR_PHASE_A = 32,

  POWER_FACTOR_PHASE_B = 33,

  POWER_FACTOR_PHASE_C = 34,

  POWER_FACTOR_SCALE_FACTOR = 35,

  EXPORTED_ENERGY = 36,

  EXPORTED_ENERGY_PHASE_A = 38,

  EXPORTED_ENERGY_PHASE_B = 40,

  EXPORTED_ENERGY_PHASE_C = 42,

  IMPORTED_ENERGY = 44,

  IMPORTED_ENERGY_PHASE_A = 46,

  IMPORTED_ENERGY_PHASE_B = 48,

  IMPORTED_ENERGY_PHASE_C = 50,

  ENERGY_SCALE_FACTOR = 52,

  EVENTS = 103,

}